Install script improvements.
Added ability to persist custom local configuration through externally referenced batch file.
Example file is provided.
For local custom file, added to gitignore.
With this to be able to test faster, added variable for selecting tasks manually, and also when mysql connection information is explicitly given, skipping that page as well.
Changed all mysql calls to use connection string variable layer to aid in customization, also involded in the above mentioned.

From these additions, primitive parameters are also added for forcing prompt or none in case a condition is otherwise met either way.

Added check to existing newpass so that script could continue when that variable was not provided so the original default one would trigger it erroneously even if not creating a mangos user.

Fixed all mysql calls for version 8 or the way I was calling the program that was having trouble setting default database name to use flag prefix which works better.

Recently also heard of a want to disable all tasks manually first, so added choice for that.

Simplified color check.